MotionCollideBehavior: Removed, but added HoverCollideBehavior instead of it.
