+++ kivy/factory_registers.py
@@ -117,8 +117,8 @@ r('DragBehavior', module='kivy.uix.behaviors.drag')
 r('EmacsBehavior', module='kivy.uix.behaviors.emacs')
 r('FocusBehavior', module='kivy.uix.behaviors.focus')
 r('HoverBehavior', module='kivy.uix.behaviors.hover')
+r('HoverCollideBehavior', module='kivy.uix.behaviors.hover')
 r('KNSpaceBehavior', module='kivy.uix.behaviors.knspace')
-r('MotionCollideBehavior', module='kivy.uix.behaviors.motion')
 r('ToggleButtonBehavior', module='kivy.uix.behaviors.togglebutton')
 r('TouchRippleBehavior', module='kivy.uix.behaviors.touchripple')
 r('TouchRippleButtonBehavior', module='kivy.uix.behaviors.touchripple')
