+++ kivy/uix/behaviors/__init__.py
@@ -87,8 +87,8 @@ __all__ = (
     'EmacsBehavior',
     'FocusBehavior',
     'HoverBehavior',
+    'HoverCollideBehavior',
     'KNSpaceBehavior',
-    'MotionCollideBehavior',
     'ToggleButtonBehavior',
     'TouchRippleBehavior',
     'TouchRippleButtonBehavior'
@@ -101,9 +101,8 @@ from kivy.uix.behaviors.cover import CoverBehavior
 from kivy.uix.behaviors.drag import DragBehavior
 from kivy.uix.behaviors.emacs import EmacsBehavior
 from kivy.uix.behaviors.focus import FocusBehavior
-from kivy.uix.behaviors.hover import HoverBehavior
+from kivy.uix.behaviors.hover import HoverBehavior, HoverCollideBehavior
 from kivy.uix.behaviors.knspace import KNSpaceBehavior
-from kivy.uix.behaviors.motion import MotionCollideBehavior
 from kivy.uix.behaviors.togglebutton import ToggleButtonBehavior
 from kivy.uix.behaviors.touchripple import TouchRippleBehavior
 from kivy.uix.behaviors.touchripple import TouchRippleButtonBehavior
